# swagger.yml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
  /tshirts:
    get:
      tags:
        - tshirt-controller
      operationId: getTshirts
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tshirt'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
    put:
      tags:
        - tshirt-controller
      operationId: updateTshirt
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Tshirt'
        required: true
      responses:
        '204':
          description: No Content
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
    post:
      tags:
        - tshirt-controller
      operationId: addTshirt
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Tshirt'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Tshirt'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
  /game/{id}:
    get:
      tags:
        - game-controller
      operationId: getBookById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Game'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
    put:
      tags:
        - game-controller
      operationId: updateGameById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Game'
        required: true
      responses:
        '202':
          description: Accepted
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Game'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
    delete:
      tags:
        - game-controller
      operationId: deleteGameById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '204':
          description: No Content
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
  /consoles:
    get:
      tags:
        - console-controller
      operationId: getAllConsoles
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Console'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
    put:
      tags:
        - console-controller
      operationId: updateConsole
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Console'
        required: true
      responses:
        '204':
          description: No Content
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
    post:
      tags:
        - console-controller
      operationId: addConsole
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Console'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Console'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
  /invoice:
    get:
      tags:
        - invoice-controller
      operationId: getAllInvoices
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InvoiceViewModel'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
    post:
      tags:
        - invoice-controller
      operationId: addInvoice
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvoiceViewModel'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/InvoiceViewModel'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
  /game:
    get:
      tags:
        - game-controller
      operationId: getAllGames
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Game'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
    post:
      tags:
        - game-controller
      operationId: addGame
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Game'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Game'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
  /tshirts/{id}:
    get:
      tags:
        - tshirt-controller
      operationId: getTshirtById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Tshirt'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
    delete:
      tags:
        - tshirt-controller
      operationId: deleteTshirt
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '204':
          description: No Content
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
  /tshirts/size/{size}:
    get:
      tags:
        - tshirt-controller
      operationId: getTshirtBySize
      parameters:
        - name: size
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tshirt'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
  /tshirts/color/{color}:
    get:
      tags:
        - tshirt-controller
      operationId: getTshirtByColor
      parameters:
        - name: color
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tshirt'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
  /invoice/{id}:
    get:
      tags:
        - invoice-controller
      operationId: getInvoiceById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/InvoiceViewModel'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
  /invoice/name/{name}:
    get:
      tags:
        - invoice-controller
      operationId: getInvoiceByCustomerName
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InvoiceViewModel'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
  /game/title/{title}:
    get:
      tags:
        - game-controller
      operationId: findGameByTitle
      parameters:
        - name: title
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Game'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
  /game/studio/{studio}:
    get:
      tags:
        - game-controller
      operationId: getAllGamesByStudio
      parameters:
        - name: studio
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Game'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
  /game/rating/{esrb}:
    get:
      tags:
        - game-controller
      operationId: getAllGamesByEsrb
      parameters:
        - name: esrb
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Game'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
  /consoles/{id}:
    get:
      tags:
        - console-controller
      operationId: getConsoleById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Console'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
    delete:
      tags:
        - console-controller
      operationId: deleteConsole
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '204':
          description: No Content
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
  /consoles/manufacturer/{manufacturer}:
    get:
      tags:
        - console-controller
      operationId: getConsolesByManufacturer
      parameters:
        - name: manufacturer
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Console'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
components:
  schemas:
    CustomErrorResponse:
      type: object
      properties:
        errorMsg:
          type: string
        status:
          type: integer
          format: int32
        errorCode:
          type: string
        timestamp:
          type: string
          format: date-time
    Tshirt:
      required:
        - color
        - description
        - price
        - quantity
        - size
      type: object
      properties:
        id:
          type: integer
          format: int32
        size:
          type: string
        color:
          type: string
        description:
          type: string
        price:
          type: number
        quantity:
          type: integer
          format: int32
    Game:
      required:
        - description
        - esrbRating
        - price
        - quantity
        - studio
        - title
      type: object
      properties:
        gameId:
          type: integer
          format: int32
        title:
          type: string
        esrbRating:
          type: string
        description:
          type: string
        price:
          type: number
        studio:
          type: string
        quantity:
          minimum: 0
          type: integer
          format: int32
    Console:
      required:
        - manufacturer
        - model
      type: object
      properties:
        id:
          type: integer
          format: int32
        model:
          type: string
        manufacturer:
          type: string
        memoryAmount:
          type: string
        processor:
          type: string
        price:
          minimum: 0
          type: number
        quantity:
          minimum: 0
          type: integer
          format: int32
    InvoiceViewModel:
      required:
        - city
        - itemId
        - itemType
        - name
        - quantity
        - state
        - street
        - zipcode
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        street:
          type: string
        city:
          type: string
        state:
          type: string
        zipcode:
          type: string
        itemType:
          type: string
        itemId:
          type: integer
          format: int32
        quantity:
          type: integer
          format: int32
        unitPrice:
          type: number
        subTotal:
          type: number
        tax:
          type: number
        processingFee:
          type: number
        total:
          type: number